requirements.txt
PyQt6
PyAutoGUI
Pillow
opencv-python
easyocr
pynput
pyinstaller
